Small refactor: Move TokenMap and TokenList to objects.{cpp,h}
